Initial Commit
This initial commit has 5 HTML files, and global CSS and JS files in their own directories. These global files are linked to in every existing HTML document.

The HTML documents contain various HTML5 semantic elements, various meta tags (though not all have content) and there are links to other pages in a <nav><ul><li><a> structure on all pages.

Anything that is in can change, this is just to get things started.