Apply .container class for Breakpoints
I've set widths for each breakpoint (ranging from just a margin each side for mobile up to set widths for larger screen sizes). I've used a class .container because there may be times we don't want to apply this to an element.

Also included some refinements to the input styling for the form page.